Homepage flexbox changes
To make the page even more responsive, I defined two main columns within a row inside a container. The image element in the second column disappears once it reaches the medium (md) breakpoint.

I also added a section with 3 main cards which also stack on one another once it reaches the medium breakpoint. 

Next: the footer section, which also needs to be responsive for small devices